requirements.txt
arrow
jsonpath_rw
psycopg2-binary
requests
lxml
pytest
pytest-cov